refactoring and small fixes

+++ commands/process.py
@@ -7,8 +7,8 @@ import pandas as pd
 
 from lib.metrics import (calculate_uniqueness_by_entity_type_and_property,
                          count_entities_by_tag, count_entities_by_type)
-from lib.schema import (append_fact, ensure_dim_date, ensure_dim_entity_type,
-                        ensure_dim_metric, ensure_dim_tags, ensure_fact)
+from lib.schema import (append_fact, merge_dim_date, merge_dim_entity_type,
+                        merge_dim_metric, merge_dim_tags, overwrite)
 
 
 def _prepare_entities(source_parquet_path: str) -> pd.DataFrame:
@@ -18,7 +18,7 @@ def _prepare_entities(source_parquet_path: str) -> pd.DataFrame:
     For example, convert string to datetime or integer, or split string into list.
     """
     entities_df = pd.read_parquet(source_parquet_path)
-    entities_df.set_index('id', inplace=True)
+    entities_df.rename(columns={'id': 'Key'}, inplace=True)
 
     entities_df['discoveryDate'] = pd.to_datetime(
         entities_df['discoveryDate']) \
@@ -55,7 +55,8 @@ def process():
     for filename in os.listdir(data_folder):
         file_path = os.path.join(data_folder, filename)
         try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
+            if filename != '.gitkeep' and \
+                    os.path.isfile(file_path) or os.path.islink(file_path):
                 os.unlink(file_path)
             elif os.path.isdir(file_path):
                 shutil.rmtree(file_path)
@@ -70,7 +71,7 @@ def process():
 
             # Fact_Entities
 
-            ensure_fact(
+            overwrite(
                 _prepare_entities(f'fixtures/{file}'),
                 'data/Fact_Entities.parquet')
 
@@ -78,16 +79,26 @@ def process():
 
             # Dim tables from Fact_Entities
 
-            ensure_dim_entity_type(
+            merge_dim_entity_type(
                 fact_entities_df,
                 'data/Dim_EntityType.parquet')
 
-            ensure_dim_tags(
+            merge_dim_tags(
                 fact_entities_df,
                 'data/Dim_Tag.parquet')
 
-            ensure_dim_date(
+            merge_dim_date(
                 'data/Dim_Date.parquet')
+
+            # Bridge_Entities_Tags
+            entities_tags_df = pd.DataFrame(
+                fact_entities_df[['Key', 'tags']].explode('tags')
+            )
+            entities_tags_df = entities_tags_df[entities_tags_df['tags'] != '']
+            entities_tags_df.columns = ['Entity_Key', 'Tag_Key']
+            overwrite(
+                entities_tags_df,
+                'data/Bridge_Entities_Tags.parquet')
 
             # Fact_DataQuality
 
@@ -110,6 +121,14 @@ def process():
             fact_data_quality_df = pd.read_parquet(
                 'data/Fact_DataQuality.parquet')
 
-            ensure_dim_metric(
+            merge_dim_metric(
                 fact_data_quality_df,
                 'data/Dim_Metric.parquet')
+
+            # dump all parquet tables in the `data/` folder to csv
+            for filename in os.listdir(data_folder):
+                file_path = os.path.join(data_folder, filename)
+                if file_path.endswith('.parquet'):
+                    df = pd.read_parquet(file_path)
+                    df.to_csv(file_path.replace(
+                        '.parquet', '.csv'), index=False)
